resume-analyzer-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
python-multipart
pypdf
google-generativeai